chore(serverless-contracts): move context generation
